correction int

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthlyIncome = float(input("Enter your monthly income: "))
-monthlyExpense = float(input("Enter your monthly expenses"))
+monthlyIncome = int(input("Enter your monthly income: "))
+monthlyExpense = int(input("Enter your monthly expenses"))
 monthlySavings = monthlyIncome-monthlyExpense
 projectedSavings = monthlySavings * 12 + (monthlySavings *12*0.05)
 print("Your monthly savings are $" + str(monthlySavings))
